requirements.txt
# UI
PyQt6>=6.6.0

# 文档解析
python-docx>=1.1.0
openpyxl>=3.1.2
PyMuPDF>=1.23.0

# LLM
openai>=1.12.0
httpx>=0.27.0

# API服务
fastapi>=0.109.0
uvicorn>=0.27.0

# 数据库
sqlalchemy>=2.0.25

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# 工具
pydantic>=2.6.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# 爬虫
beautifulsoup4>=4.12.0
lxml>=5.1.0
